# packages/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "App": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/App.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "auth": {
            "$ref": "#/components/schemas/AuthConfig"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "oauth_client_id": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          }
        },
        "required": ["id", "name", "hostname", "created_at"],
        "type": "object"
      },
...
      "AuthConfig": {
        "additionalProperties": false,
        "properties": {
          "protected_routes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          }
        },
        "type": "object"
      },
...
      "CreateAppInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/CreateAppInput.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "auth": {
            "$ref": "#/components/schemas/AuthConfig"
          },
          "external_id": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name", "hostname"],
        "type": "object"
      },
...
      "OpenAPIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OpenAPIError.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": ["error"],
        "type": "object"
      },
...
      "WorkerDetail": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/WorkerDetail.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "app": {
            "$ref": "#/components/schemas/App"
          },
          "deployment": {
            "$ref": "#/components/schemas/WorkerDeployment"
          },
          "secrets": {
            "items": {
              "$ref": "#/components/schemas/Secret"
            },
            "type": ["array", "null"]
          }
        },
        "required": ["app"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT or personal access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nanoflare Control API",
    "version": "v1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/organizations/{orgID}/workers": {
      "get": {
        "operationId": "listWorkers",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/App"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listWorkers",
        "tags": ["Workers"]
      },
      "post": {
        "operationId": "createWorker",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/App"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "createWorker",
        "tags": ["Workers"]
      }
    },
    "/v1/organizations/{orgID}/workers/{workerID}": {
      "delete": {
        "operationId": "deleteWorker",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "deleteWorker",
        "tags": ["Workers"]
      },
      "get": {
        "operationId": "getWorker",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerDetail"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "getWorker",
        "tags": ["Workers"]
      },
      "patch": {
        "operationId": "updateWorker",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/App"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "updateWorker",
        "tags": ["Workers"]
      }
    },
    "/v1/organizations/{orgID}/workers/{workerID}/analytics/traffic": {
      "get": {
        "operationId": "getWorkerTraffic",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerTraffic"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "getWorkerTraffic",
        "tags": ["Workers"]
      }
    },
    "/v1/organizations/{orgID}/workers/{workerID}/deployment-traffic": {
      "put": {
        "operationId": "setWorkerDeploymentTraffic",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentTrafficRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsoleDeployment"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "setWorkerDeploymentTraffic",
        "tags": ["Workers"]
      }
    },
...
    "/v1/organizations/{orgID}/workers/{workerID}/files": {
      "get": {
        "operationId": "listWorkerFiles",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workerID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkerFile"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listWorkerFiles",
        "tags": ["Workers"]
      }
    },
...
  }
}